build(core): Flyway outbox migration and restore booking partial-index IT
- V5__create_event_publication_table.sql: Spring Modulith outbox table
  with schema matching Hibernate's @Entity expectations so ddl-auto:validate
  passes in prod. Unblocks @ApplicationModuleListener for at-least-once
  delivery + republication via /actuator/modulith/events.
- V4 fix: response_status SMALLINT -> INTEGER (entity uses Integer; was
  rejected by Hibernate schema validation).
- JpaBookingRepositoryIT restored: @SpringBootTest with ddl-auto: validate
  so Flyway owns the schema and the partial unique index from V3 is
  preserved (Hibernate's default create-drop would otherwise regenerate
  the table without the WHERE clause).
- memory.md: Phase D, E, F sections documenting decisions and DoD.

Tests: 101 surefire + 7 IT = 108 green (JpaBookingRepositoryIT +2).

Refs: Phase F

+++ src/main/resources/db/migration/V4__create_idempotency_keys.sql
@@ -11,7 +11,7 @@ CREATE TABLE idempotency_keys (
     user_id           UUID         NOT NULL,
     endpoint          VARCHAR(255) NOT NULL,
     request_hash      VARCHAR(64)  NOT NULL,
-    response_status   SMALLINT,
+    response_status   INTEGER,
     response_body     TEXT,
     response_type     VARCHAR(255),
     created_at        TIMESTAMPTZ  NOT NULL,
